# docs/swagger.yaml
openapi: 3.0.0
paths:
...
  /auth/signup:
    post:
      operationId: AuthController_signup
      summary: ''
      description: ''
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpInput'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: object
      tags: &ref_3
        - auth
      security: &ref_4
        - bearer: []
  /auth/login:
    post:
      operationId: AuthController_login
      summary: ''
      description: ''
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: object
      tags: *ref_3
      security: *ref_4
  /auth/profile:
    get:
      operationId: AuthController_getProfile
      summary: ''
      description: ''
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntity'
      tags: *ref_3
      security: *ref_4
  /auth/updatePassword:
    post:
      operationId: AuthController_updatePassword
      summary: ''
      description: ''
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordInput'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: object
      tags: *ref_3
      security: *ref_4
...
info:
  title: nest-api
  description: The nest API description
  version: '0.1'
  contact: {}
...
servers: []
components:
  securitySchemes:
    basic:
      type: http
      scheme: basic
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    SignUpInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    UserEntity:
      type: object
      properties:
        id:
          type: number
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - createdAt
        - updatedAt
    UpdatePasswordInput:
      type: object
      properties: {}
...